Add `Stats` component, refactor `Dashboard` for modularity, optimize transaction filtering with IDs, and introduce constants file for reusable data.
